Implemented fast loader.
